+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2018-10-17
+### Fixed
+- Fix subrouters appending middlewares to the parent's middleware slice.
+
+### Removed
+- `(*Router).SetCtxKey` in order to set the key only once, when creating a new router.
+
 ## [0.3.2] - 2018-10-03
 ### Fixed
 - Copy placeholder when creating a subrouter.
@@ -51,6 +58,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - EditorConfig file.
 - This package's source code, including examples and tests.
 
+[0.4.0]: https://github.com/gbrlsnchs/mux/compare/v0.3.2...v0.4.0
 [0.3.2]: https://github.com/gbrlsnchs/mux/compare/v0.3.1...v0.3.2
 [0.3.1]: https://github.com/gbrlsnchs/mux/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/gbrlsnchs/mux/compare/v0.2.0...v0.3.0